requirements.txt
 os
 tkinter 
 threading
 shutil
 base64
 io
 pillow
 ollama
 tqdm